新增截图功能 优化音量调节

+++ SSPlayer/function.cpp
@@ -5,6 +5,6 @@ bool thread_pause = false;
 bool isMute = false;
 
 double fps = 25;
-int silence = SDL_MIX_MAXVOLUME / 2;
+double silence = SDL_MIX_MAXVOLUME / 2;
 
 uint8_t *copy_buf;
